LPS-54497 Move wiki portlet instance settings to wiki-web module because wiki-api is meant to be the API of wiki-service not of the whole wiki application.
