+++ modules/apps/wiki/wiki-web/docroot/WEB-INF/src/com/liferay/wiki/web/settings/definition/WikiPortletInstanceSettingsExtraImpl.java
@@ -12,13 +12,13 @@
  * details.
  */
 
-package com.liferay.wiki.settings.definition;
+package com.liferay.wiki.web.settings.definition;
 
 import com.liferay.portal.kernel.settings.ModifiableSettings;
 import com.liferay.portal.kernel.settings.Settings;
 import com.liferay.portal.kernel.settings.TypedSettings;
 import com.liferay.portal.util.PortalUtil;
-import com.liferay.wiki.settings.WikiPortletInstanceSettingsExtra;
+import com.liferay.wiki.web.settings.WikiPortletInstanceSettingsExtra;
 
 import java.io.IOException;
