+++ modules/apps/wiki/wiki-web/docroot/html/portlet/wiki/init.jsp
@@ -48,7 +48,6 @@ page import="com.liferay.wiki.service.permission.WikiNodePermission" %><%@
 page import="com.liferay.wiki.service.permission.WikiPagePermission" %><%@
 page import="com.liferay.wiki.service.permission.WikiPermission" %><%@
 page import="com.liferay.wiki.settings.WikiGroupServiceSettings" %><%@
-page import="com.liferay.wiki.settings.WikiPortletInstanceSettings" %><%@
 page import="com.liferay.wiki.social.WikiActivityKeys" %><%@
 page import="com.liferay.wiki.util.WikiCacheUtil" %><%@
 page import="com.liferay.wiki.util.WikiPageAttachmentsUtil" %><%@
@@ -59,6 +58,7 @@ page import="com.liferay.wiki.web.display.context.logic.WikiPortletInstanceSetti
 page import="com.liferay.wiki.web.display.context.logic.WikiVisualizationHelper" %><%@
 page import="com.liferay.wiki.web.display.context.util.WikiRequestHelper" %><%@
 page import="com.liferay.wiki.web.display.context.util.WikiURLHelper" %><%@
+page import="com.liferay.wiki.web.settings.WikiPortletInstanceSettings" %><%@
 page import="com.liferay.wiki.web.settings.WikiWebSettingsProvider" %>
 
 <%
